project/pyproject.toml
[tool.poetry]
name = "LINE Bot Tutorial"
version = "0.1.0"
description = "FastAPI Example for LINE Bot"
authors = ["FawenYo <twsky020600@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.6"
fastapi = "^0.63.0"
line-bot-sdk = "^1.19.0"
uvicorn = "^0.13.4"
python-dotenv = "^0.17.0"
aiofiles = "^0.6.0"
gunicorn = "^20.1.0"
Jinja2 = "^2.11.3"
uvloop = "^0.15.2"
httptools = "^0.1.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
